fix(spec): derive generated-PR delivery IDs and append subject keys
:402 — response.result.generated_pr_delivery_id was only substituted into
the positive fixtures, so a caller-supplied literal passed through with
recomputed result/response digests. It is now recomputed from the request,
and an attacker-chosen value fails the gate.

:329 — the ledger contract makes expected_attempt_subject_key a byte-equal
assertion over the authority-derived subject, but the wire carried only a
by-reference ledger_object, so no conformance gate could recompute it. The
request body now carries the closed attempt_subject payload the contract
already defines field-by-field, with a matching attempt_subject_key digest
node, runtime domain, and DAG edge. The assertion is excluded from the
operation_id subject, because the contract says the caller's key must not be
an authority input.

Verified: substituting the key, and changing the payload while pinning the
key to the old subject, both fail.

verify_publication_authority_api.py crossed the U-16 ceiling, so the Draft
2020-12 evaluator moved to publication_schema_validator.py; make_validator
injects the caller's error type, uint64 validator, and canonicalizer so the
module carries no verifier-specific imports.
