+++ docs/specs/GH700/publication_authority_api.models.json
@@ -51,6 +51,7 @@
       "schema_version":"GH700:initial-time-proof-bundle:v1", "release_identity_attestation_digest":{"$fixture":"digest"},
       "trusted_lower_bound":1700000000, "trusted_upper_bound":1800000000, "trusted_time_proof_digest":{"$fixture":"digest"}
     },
+    "attempt_subject": { "attempt_record_kind":"candidate_failure", "candidate_identity_or_null":{"$fixture":"digest"}, "failure_scope_or_null":"build", "target_or_null":"release", "early_attempt_key_or_null":null },
     "ledger_object": {
       "schema_version": "GH700:ledger-object-ref:v1",
       "object_kind": "blocked_attempt_record",
@@ -378,7 +379,7 @@
     { "model_id":"client.plan_generated_pr", "surface":"client", "method":"plan_generated_pr", "request_base":"client_request_base", "request_patch":{"method":"plan_generated_pr","expected_publication_frontier_or_null":{"$fixture":"publication_frontier"},"body":{"intent":{"$fixture":"history_object"},"append_authorization":{"$fixture":"append_auth_plan_pr"}}}, "success_base":"client_success_base", "success_patch":{"method":"plan_generated_pr","result":{"planned_transition_receipt":{"$fixture":"transition_receipt"},"generated_pr_delivery_id":{"$context":"generated_pr_delivery_id"}}} },
     { "model_id":"client.deliver_generated_pr", "surface":"client", "method":"deliver_generated_pr", "request_base":"client_request_base", "request_patch":{"method":"deliver_generated_pr","expected_publication_frontier_or_null":{"$fixture":"publication_frontier"},"body":{"planned_operation_id":{"$fixture":"digest"},"generated_pr_delivery_id":{"$derive":"generated_pr_delivery_id"},"delivery_authorization":{"$fixture":"delivery_auth_pr"}}}, "success_base":"client_success_base", "success_patch":{"method":"deliver_generated_pr","result":{"delivery_state":"bound","transition_receipt":{"$fixture":"generated_pr_bound_receipt"},"send_once_audit_digest":{"$fixture":"digest"}}} },
     { "model_id":"client.recover_generated_pr", "surface":"client", "method":"recover_generated_pr", "request_base":"client_request_base", "request_patch":{"method":"recover_generated_pr","expected_publication_frontier_or_null":{"$fixture":"publication_frontier"},"body":{"planned_operation_id":{"$fixture":"digest"},"recovery_query_digest":{"$derive":"recovery_query_digest"},"append_authorization":{"$fixture":"append_auth_recover_pr"}}}, "success_base":"client_success_base", "success_patch":{"method":"recover_generated_pr","result":{"recovery_state":"bound_existing","transition_receipts":[{"$fixture":"generated_pr_bound_receipt"}]}} },
-    { "model_id":"client.append_blocked_attempt", "surface":"client", "method":"append_blocked_attempt", "request_base":"client_request_base", "request_patch":{"method":"append_blocked_attempt","expected_blocked_attempt_frontier_or_null":{"$fixture":"blocked_frontier"},"body":{"attempt_record":{"$fixture":"ledger_object"},"expected_attempt_subject_key":{"$fixture":"digest"},"ledger_append_authorization":{"$fixture":"ledger_authorization"}}}, "success_base":"client_success_base", "success_patch":{"method":"append_blocked_attempt","result":{"attempt_record_receipt":{"$fixture":"ledger_receipt"},"blocked_attempt_frontier_receipt":{"$fixture":"blocked_frontier_receipt"}}} },
+    { "model_id":"client.append_blocked_attempt", "surface":"client", "method":"append_blocked_attempt", "request_base":"client_request_base", "request_patch":{"method":"append_blocked_attempt","expected_blocked_attempt_frontier_or_null":{"$fixture":"blocked_frontier"},"body":{"attempt_record":{"$fixture":"ledger_object"},"attempt_subject":{"$fixture":"attempt_subject"},"expected_attempt_subject_key":{"$derive":"expected_attempt_subject_key"},"ledger_append_authorization":{"$fixture":"ledger_authorization"}}}, "success_base":"client_success_base", "success_patch":{"method":"append_blocked_attempt","result":{"attempt_record_receipt":{"$fixture":"ledger_receipt"},"blocked_attempt_frontier_receipt":{"$fixture":"blocked_frontier_receipt"}}} },
     { "model_id":"client.bind_blocked_attempt", "surface":"client", "method":"bind_blocked_attempt", "request_base":"client_request_base", "request_patch":{"method":"bind_blocked_attempt","expected_publication_frontier_or_null":{"$fixture":"publication_frontier"},"expected_blocked_attempt_frontier_or_null":{"$fixture":"blocked_frontier"},"body":{"binding_intent":{"$fixture":"terminal_binding"},"ledger_append_authorization":{"$fixture":"ledger_auth_bind"}}}, "success_base":"client_success_base", "success_patch":{"method":"bind_blocked_attempt","result":{"binding_record_receipt":{"$fixture":"ledger_receipt_bound"},"blocked_attempt_frontier_receipt":{"$fixture":"blocked_frontier_receipt"}}}, "coverage_tags":["terminal_binding"] },
     { "model_id":"client.list_blocked_attempts", "surface":"client", "method":"list_blocked_attempts", "request_base":"client_request_base", "request_patch":{"method":"list_blocked_attempts","expected_blocked_attempt_frontier_or_null":{"$fixture":"blocked_frontier"},"body":{"source_identity_key":{"$fixture":"digest"},"candidate_identity_or_null":null,"run_id_or_null":null,"run_attempt_or_null":null,"attempt_record_kind_or_null":null,"attempt_subject_key_or_null":null,"page_cursor_or_null":null,"page_size":50}}, "success_base":"client_success_base", "success_patch":{"method":"list_blocked_attempts","result":{"attempt_records":[],"next_page_cursor_or_null":null,"enumeration_snapshot_receipt":{"$fixture":"enumeration_receipt"}}} },
     { "model_id":"client.commit_reconciliation_watermark", "surface":"client", "method":"commit_reconciliation_watermark", "request_base":"client_request_base", "request_patch":{"method":"commit_reconciliation_watermark","expected_blocked_attempt_frontier_or_null":{"$fixture":"blocked_frontier"},"body":{"reconciliation_watermark":{"$fixture":"ledger_object"},"terminal_listing_proof":{"$fixture":"ledger_object"},"ledger_append_authorization":{"$fixture":"ledger_auth_watermark"}}}, "success_base":"client_success_base", "success_patch":{"method":"commit_reconciliation_watermark","result":{"watermark_receipt":{"$fixture":"ledger_receipt_watermark"},"blocked_attempt_frontier_receipt":{"$fixture":"blocked_frontier_receipt"}}} },
